requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
redis==4.5.1
sqlparse==0.4.3
